requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
guardrails-ai==0.5.0
google-generativeai==0.3.2
pydantic==2.8.2
python-dotenv==1.0.0
pytest==7.4.3
python-multipart==0.0.9
